+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-- [feature] Add `EPTSDK.PaymentSubscriptions.confirm/3` for the new intent-to-subscription confirm flow
+- [feature] Add `EPTSDK.PaymentSubscription.confirm/3` for the new intent-to-subscription confirm flow
 
 ## 13.0.2